eol-globi-resource-local/src/main/java/org/eol/globi/util/ResourceUtil.java
package org.eol.globi.util;

import org.apache.commons.lang3.StringUtils;

import java.net.URI;

public class ResourceUtil {

    public static boolean isFileURI(URI resource) {
        return StringUtils.startsWith(resource.getScheme(), "file");
    }

    public static URI getAbsoluteResourceURI(URI context, URI resourceName) {
        return resourceName.isAbsolute()
                ? resourceName
                : absoluteURIFor(context, resourceName);
    }


    private static URI absoluteURIFor(URI context, URI resourceName) {
        String resourceNameNoSlashPrefix = StringUtils.startsWith(resourceName.toString(), "/")
                ? StringUtils.substring(resourceName.toString(), 1)
                : resourceName.toString();
        String contextString = context.toString();
        String contextNoSlashSuffix = StringUtils.endsWith(contextString, "/")
                ? StringUtils.substring(contextString, 0, contextString.length() - 1)
                : contextString;

        return URI.create(contextNoSlashSuffix + "/" + resourceNameNoSlashPrefix);
    }


}
